[xen] Stubdoms map_pirq/unmap_pirq.
Passed through devices with MSI capabilities might require QEMU to
map/unmap pirq. Then only dom0 is granted access to those pirq's irq
which cause EPERM when QEMU does so from the stubdomain.

Check if the domain, for which the pirq is mapped, is the target of the
one issuing the hypercall, then give it access to that irq (and deny it
in due time).

OXT-131.